userChrome: give the DOM 100ms to catch up with touch events before closing OSK
Ever since we switched to phosh-osk-stub, tapping on an icon in the start page
would result in the touch event getting processed for the icon on top of the
icon the user actually pressed. This is probably happening because squeekboard
might have taken a bit longer to close than phosh-osk-stub does, so by the time
the finger was raised and the event was processed, the page had reflowed.

100ms seems like the sweet spot - hover states look a bit wonky, but the correct
page gets loaded, and the search bar works.

This is just a workaround, I believe the keyboard itself should debounce open and
close events. But it fixes the immediate problem.