debian: drop 2&>1 and set -e from postinst
it turns out, 2&>1 changes the return code in dash, but not bash. and again it turns out, 2&>1 also masks failures from set -e! drop them both.
so this script was actually broken this whole time, but it was broken in a way that made it functional for us. kudos to myself for breaking, and then fixing it after over 6 months and never noticing a thing.

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>